+++ packages/node/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Getting transaction receipts when accessing via a log handler
+- transaction missing from log (#202)
 
 ## [3.3.0] - 2023-11-06
 ### Added